fixing-types

+++ copula/birth.py
@@ -20,7 +20,7 @@ def birth(currentModel, u, dist, numbrk, q):
         bir = np.sort(new)
         j2 = np.count_nonzero(new == 0)
         d = np.argwhere(bir == kn) + 1
-        d = int(d)
+        d = d
         t2 = currentModel[np.sort(currentModel) != 0]
 
         if kn > np.max(t2):
@@ -52,7 +52,7 @@ def birth(currentModel, u, dist, numbrk, q):
                     if a >= 2.0/3:
                         Q[numbrk - j2 - 1:numbrk - j2 + 1] = q[numbrk - j2 - 1]
 
-            Q[d - j2+1:numbrk + 1] = Q[d - j2] * np.ones(numbrk - d + j2)
+            Q[d - j2+1:numbrk + 1] = Q[d - j2] * np.ones(numbrk - d + j2, dtype=int)
             s = np.concatenate((np.asarray([q[numbrk - j2 - 1]]), np.asarray(Q[numbrk - j2 - 1:numbrk - j2 + 1])), axis=0)
         else:
             if kn < np.min(t2):
@@ -79,7 +79,7 @@ def birth(currentModel, u, dist, numbrk, q):
                         Q[:2] = list(temp[row - 1])[::-1]
                     else:
                         if a >= 2.0/3:
-                            Q[:2] = np.ones(2) * q[0]
+                            Q[:2] = np.ones(2,dtype=int) * q[0]
 
                 s = np.concatenate((np.asarray([q[0]]), Q[:2]), axis=0)
 
@@ -107,7 +107,7 @@ def birth(currentModel, u, dist, numbrk, q):
                         Q[d - j2 - 1:d - j2 + 1] = list(temp[row - 1])[::-1]
                     else:
                         if a >= 2.0/3:
-                            Q[d - j2 - 1:d - j2 + 1] = np.ones(2) * q[d - j2 - 1]
+                            Q[d - j2 - 1:d - j2 + 1] = np.ones(2,dtype=int) * q[d - j2 - 1]
 
                 Q[d - j2 + 1: numbrk + 1] = q[d - j2:numbrk]
                 s = np.concatenate((np.asarray([q[d - j2 - 1]]), Q[d - j2 - 1:d - j2 + 1]), axis=0)
@@ -120,7 +120,7 @@ def birth(currentModel, u, dist, numbrk, q):
         temp = [comb for comb in combinations([1, 2, 3], 2)]
         j = 0
         G = 0
-        Q = np.ones(numbrk + 1)
+        Q = np.ones(numbrk + 1,dtype=int)
 
         while j < 3 and G == 0:
             if np.all(temp[j] != np.ones(shape=(1,2)) * q[0]):
