+++ copula/kill.py
@@ -18,13 +18,13 @@ def kill(currentModel, numbrk, q):
 
     if(Kn >= numbrk and L != 1):
         if(q[kn - 1] != q[kn]):
-            if(w <= 1.0/2):
-                Q[kn - 1: numbrk + 1] = q[kn] * np.ones(shape=(numbrk + 2 - kn))
+            if(w <= 1/2):
+                Q[kn - 1: numbrk + 1] = q[kn] * np.ones(shape=(numbrk + 2 - kn),dtype=int)
             else:
-                Q[kn - 1: numbrk + 1] = q[kn - 1] * np.ones(shape=(numbrk + 2 - kn))
+                Q[kn - 1: numbrk + 1] = q[kn - 1] * np.ones(shape=(numbrk + 2 - kn),dtype=int)
 
         else:
-            Q[kn - 1: numbrk + 1] = q[kn - 1] * np.ones(shape=(numbrk + 2 - kn))
+            Q[kn - 1: numbrk + 1] = q[kn - 1] * np.ones(shape=(numbrk + 2 - kn),dtype=int)
 
         s = np.concatenate((np.asarray([q[kn - 1]]), np.asarray([q[kn]]), np.asarray([Q[kn]])), axis=0)
 
@@ -33,10 +33,10 @@ def kill(currentModel, numbrk, q):
             if q[0] != q[1]:
                 Q = q[:]
             else:
-                if(w <= 1.0/2):
-                    Q = q[0] * np.ones(shape=(numbrk + 1))
+                if(w <= 1/2):
+                    Q = q[0] * np.ones(shape=(numbrk + 1),dtype=int)
                 else:
-                    Q = q[1] * np.ones(shape=(numbrk + 1))
+                    Q = q[1] * np.ones(shape=(numbrk + 1),dtype=int)
 
             s = np.concatenate((np.asarray([q[0]]), np.asarray([q[1]]), np.asarray([Q[0]])), axis=0)
 
@@ -44,7 +44,7 @@ def kill(currentModel, numbrk, q):
             Q = np.delete(Q, kn - 1)
 
             if q[kn] != q[kn - 1]:
-                if w <= 1.0/2:
+                if w <= 1/2:
                     Q[kn - 1] = q[kn]
                 else:
                     Q[kn - 1] = q[kn - 1]
